begin initial transition of tests to libarmcc
	modified:   libarmcc/source/include/armcc_gen_arm_dp_sop.h
	deleted:    libarmvm/test/include/arm_gen_ir.hpp
	deleted:    libarmvm/test/include/arm_gen_sop.hpp
	deleted:    libarmvm/test/include/test.h
	new file:   libarmvm/test/include/test.hpp
	deleted:    libarmvm/test/include/test_arm.h
	deleted:    libarmvm/test/include/test_arm_gen.hpp
	deleted:    libarmvm/test/include/test_gen_utility.h
	modified:   libarmvm/test/makefile
	new file:   libarmvm/test/source/include/test_arm.hpp
	renamed:    libarmvm/test/include/test_armvm_mem.h -> libarmvm/test/source/include/test_armvm_mem.h
	renamed:    libarmvm/test/include/test_utility.h -> libarmvm/test/source/include/test_utility.hpp
	renamed:    libarmvm/test/source/test.c -> libarmvm/test/source/test.cpp
	deleted:    libarmvm/test/source/test_arm.c
	new file:   libarmvm/test/source/test_arm.cpp
	renamed:    libarmvm/test/source/test_armvm_mem.c -> libarmvm/test/source/test_armvm_mem.cx

+++ libarmcc/source/include/armcc_gen_arm_dp_sop.h
@@ -14,8 +14,8 @@
 
 /* **** */
 
-static armcc_sop_t gen_arm_dp_sop__rm_i(const arm_reg_t rm, const arm_sop_t shift_type, const uint8_t rs)
+static armcc_sop_t gen_arm_dp_sop__rm_i(const arm_reg_t rm, const arm_sop_t shift_type, const int rs)
 { return(pbBFMOV(rs, 0, 5, 7) | gen_arm_ir__shift_type(shift_type) | gen_arm_ir__rm(rm)); }
 
-static armcc_sop_t gen_arm_dp_sop__ror_i(const uint8_t imm, const uint8_t rs)
+static armcc_sop_t gen_arm_dp_sop__ror_i(const int imm, const int rs)
 { return((imm & 0xff) | pbBFMOV(rs, 1, 4, 8)); }
